@@ -73,23 +76,24 @@
     "babel-preset-stage-3": "6.24.1",
     "babel-register": "6.26.0",
     "babel-runtime": "6.26.0",
-    "codecov": "3.0.2",
+    "bundlesize": "0.17.0",
     "cross-env": "5.1.4",
+    "codecov": "3.0.4",
     "css-loader": "0.28.10",
     "element-react": "1.4.8",
     "element-theme-default": "1.4.13",
     "enzyme": "3.3.0",
     "enzyme-adapter-react-16": "1.1.1",
-    "eslint": "4.18.2",
+    "eslint": "5.2.0",
     "eslint-config-google": "0.9.1",
-    "eslint-loader": "2.0.0",
-    "eslint-plugin-babel": "4.1.2",
-    "eslint-plugin-flowtype": "2.46.1",
-    "eslint-plugin-import": "2.9.0",
-    "eslint-plugin-jest": "21.14.0",
-    "eslint-plugin-react": "7.7.0",
+    "eslint-loader": "2.1.0",
+    "eslint-plugin-babel": "5.1.0",
+    "eslint-plugin-flowtype": "2.50.0",
+    "eslint-plugin-import": "2.13.0",
+    "eslint-plugin-jest": "21.18.0",
+    "eslint-plugin-react": "7.10.0",
     "file-loader": "1.1.11",
-    "flow-bin": "0.69.0",
+    "flow-bin": "0.77.0",
     "flow-runtime": "0.17.0",
     "friendly-errors-webpack-plugin": "1.7.0",
     "github-markdown-css": "2.10.0",
@@ -97,41 +101,43 @@
     "husky": "0.15.0-rc.8",
     "identity-obj-proxy": "3.0.0",
     "in-publish": "2.0.0",
-    "jest": "22.4.3",
-    "jest-environment-jsdom": "22.4.3",
-    "jest-environment-jsdom-global": "1.0.3",
-    "jest-environment-node": "22.4.3",
+    "jest": "23.4.1",
+    "jest-environment-jsdom": "23.4.0",
+    "jest-environment-jsdom-global": "1.1.0",
+    "jest-environment-node": "23.4.0",
     "localstorage-memory": "1.0.2",
-    "mini-css-extract-plugin": "0.4.0",
-    "node-mocks-http": "1.6.7",
-    "node-sass": "4.9.0",
+    "mini-css-extract-plugin": "0.4.1",
+    "node-mocks-http": "1.7.0",
+    "node-sass": "4.9.2",
     "normalize.css": "8.0.0",
+    "optimize-css-assets-webpack-plugin": "5.0.0",
     "ora": "1.4.0",
     "prop-types": "15.6.1",
     "puppeteer": "1.1.1",
     "react": "16.2.0",
     "react-dom": "16.2.0",
-    "react-hot-loader": "4.0.0",
+    "react-hot-loader": "4.2.0",
     "react-router-dom": "4.2.2",
     "react-syntax-highlighter": "5.8.0",
     "rimraf": "2.6.2",
     "sass-loader": "6.0.7",
     "source-map-loader": "0.2.3",
-    "standard-version": "4.3.0",
-    "style-loader": "0.20.3",
-    "stylelint": "9.1.1",
-    "stylelint-config-recommended-scss": "3.1.0",
+    "standard-version": "4.4.0",
+    "style-loader": "0.21.0",
+    "stylelint": "9.4.0",
+    "stylelint-config-recommended-scss": "3.2.0",
     "stylelint-scss": "2.5.0",
-    "stylelint-webpack-plugin": "0.10.4",
-    "supertest": "3.0.0",
+    "stylelint-webpack-plugin": "0.10.5",
+    "supertest": "3.1.0",
     "url-loader": "0.6.2",
     "verdaccio-auth-memory": "0.0.4",
-    "verdaccio-memory": "1.0.1",
-    "webpack": "4.8.3",
-    "webpack-cli": "2.0.15",
-    "webpack-dev-server": "3.1.4",
-    "webpack-merge": "4.1.2",
-    "whatwg-fetch": "2.0.3"
+    "verdaccio-memory": "1.0.3",
+    "webpack": "4.16.3",
+    "webpack-bundle-analyzer": "2.13.1",
+    "webpack-cli": "3.1.0",
+    "webpack-dev-server": "3.1.5",
+    "webpack-merge": "4.1.3",
+    "whatwg-fetch": "2.0.4"
   },
   "keywords": [
     "private",
