@@ -4,6 +4,7 @@ import _ from 'lodash';
 import Path from 'path';
 import mime from 'mime';
 
+import {API_MESSAGE, HEADERS} from '../../../lib/constants';
 import {DIST_TAGS, validate_metadata, isObject, ErrorCode} from '../../../lib/utils';
 import {media, expectJson, allow} from '../../middleware';
 import {notify} from '../../../lib/notify';
@@ -11,6 +12,7 @@ import {notify} from '../../../lib/notify';
 import type {Router} from 'express';
 import type {Config, Callback} from '@verdaccio/types';
 import type {IAuth, $ResponseExtend, $RequestExtend, $NextFunctionVer, IStorageHandler} from '../../../../types';
+import logger from '../../../lib/logger';
 
 export default function(router: Router, auth: IAuth, storage: IStorageHandler, config: Config) {
   const can = allow(auth);
@@ -65,7 +67,7 @@ export default function(router: Router, auth: IAuth, storage: IStorageHandler, c
         || Object.keys(metadata.versions).length !== 1) {
         // npm is doing something strange again
         // if this happens in normal circumstances, report it as a bug
-        return next(ErrorCode.get400('unsupported registry call'));
+        return next(ErrorCode.getBadRequest('unsupported registry call'));
       }
 
       if (err && err.status != 409) {
@@ -79,18 +81,24 @@ export default function(router: Router, auth: IAuth, storage: IStorageHandler, c
           return next(err);
         }
 
-        const t2 = Object.keys(metadata.versions)[0];
-        metadata.versions[t2].readme = _.isNil(metadata.readme) === false ? String(metadata.readme) : '';
-        create_version(t2, metadata.versions[t2], function(err) {
+        const versionToPublish = Object.keys(metadata.versions)[0];
+        metadata.versions[versionToPublish].readme = _.isNil(metadata.readme) === false ? String(metadata.readme) : '';
+        create_version(versionToPublish, metadata.versions[versionToPublish], function(err) {
           if (err) {
             return next(err);
           }
 
-          add_tags(metadata[DIST_TAGS], function(err) {
+          add_tags(metadata[DIST_TAGS], async function(err) {
             if (err) {
               return next(err);
             }
-            notify(metadata, config);
+
+            try {
+              await notify(metadata, config, req.remote_user, `${metadata.name}@${versionToPublish}`);
+            } catch (err) {
+              logger.logger.error({err}, 'notify batch service has failed: @{err}');
+            }
+
             res.status(201);
             return next({ok: ok_message, success: true});
           });
@@ -100,22 +108,22 @@ export default function(router: Router, auth: IAuth, storage: IStorageHandler, c
 
     if (Object.keys(req.body).length === 1 && isObject(req.body.users)) {
       // 501 status is more meaningful, but npm doesn't show error message for 5xx
-      return next(ErrorCode.get404('npm star|unstar calls are not implemented'));
+      return next(ErrorCode.getNotFound('npm star|unstar calls are not implemented'));
     }
 
     try {
       metadata = validate_metadata(req.body, name);
     } catch (err) {
-      return next(ErrorCode.get422('bad incoming package data'));
+      return next(ErrorCode.getBadData('bad incoming package data'));
     }
 
     if (req.params._rev) {
       storage.changePackage(name, metadata, req.params.revision, function(err) {
-        after_change(err, 'package changed');
+        after_change(err, API_MESSAGE.PKG_CHANGED);
       });
     } else {
       storage.addPackage(name, metadata, function(err) {
-        after_change(err, 'created new package');
+        after_change(err, API_MESSAGE.PKG_CREATED);
       });
     }
   });
@@ -127,7 +135,7 @@ export default function(router: Router, auth: IAuth, storage: IStorageHandler, c
         return next(err);
       }
       res.status(201);
-      return next({ok: 'package removed'});
+      return next({ok: API_MESSAGE.PKG_REMOVED});
     });
   });
 
@@ -139,12 +147,12 @@ export default function(router: Router, auth: IAuth, storage: IStorageHandler, c
         return next(err);
       }
       res.status(201);
-      return next({ok: 'tarball removed'});
+      return next({ok: API_MESSAGE.TARBALL_REMOVED});
     });
   });
 
   // uploading package tarball
-  router.put('/:package/-/:filename/*', can('publish'), media('application/octet-stream'),
+  router.put('/:package/-/:filename/*', can('publish'), media(HEADERS.OCTET_STREAM),
   function(req: $RequestExtend, res: $ResponseExtend, next: $NextFunctionVer) {
     const name = req.params.package;
     const stream = storage.addTarball(name, req.params.filename);
@@ -188,7 +196,7 @@ export default function(router: Router, auth: IAuth, storage: IStorageHandler, c
 
       res.status(201);
       return next({
-        ok: 'package published',
+        ok: API_MESSAGE.PKG_PUBLISHED,
       });
     });
   });
