@@ -1,13 +1,13 @@
 // @flow
 
-import crypto from 'crypto';
 import _ from 'lodash';
 import {
-  validate_name as utilValidateName,
+  validateName as utilValidateName,
   validate_package as utilValidatePackage,
   isObject,
   ErrorCode} from '../lib/utils';
-import {HEADERS} from '../lib/constants';
+import {API_ERROR, HEADER_TYPE, HEADERS, HTTP_STATUS, TOKEN_BASIC, TOKEN_BEARER} from '../lib/constants';
+import {stringToMD5} from '../lib/crypto-utils';
 import type {$ResponseExtend, $RequestExtend, $NextFunctionVer, IAuth} from '../../types';
 import type {Config} from '@verdaccio/types';
 
@@ -38,7 +38,7 @@ export function validateName(req: $RequestExtend, res: $ResponseExtend, next: $N
   } else if (utilValidateName(value)) {
     next();
   } else {
-    next( ErrorCode.get403('invalid ' + name));
+    next( ErrorCode.getForbidden('invalid ' + name));
   }
 }
 
@@ -51,15 +51,15 @@ export function validatePackage(req: $RequestExtend, res: $ResponseExtend, next:
   } else if (utilValidatePackage(value)) {
     next();
   } else {
-    next( ErrorCode.get403('invalid ' + name) );
+    next( ErrorCode.getForbidden('invalid ' + name) );
   }
 }
 
 export function media(expect: string) {
   return function(req: $RequestExtend, res: $ResponseExtend, next: $NextFunctionVer) {
-    if (req.headers['content-type'] !== expect) {
-      next( ErrorCode.getCode(415, 'wrong content-type, expect: ' + expect
-        + ', got: '+req.headers['content-type']) );
+    if (req.headers[HEADER_TYPE.CONTENT_TYPE] !== expect) {
+      next( ErrorCode.getCode(HTTP_STATUS.UNSUPORTED_MEDIA, 'wrong content-type, expect: ' + expect
+        + ', got: '+req.headers[HEADER_TYPE.CONTENT_TYPE]) );
     } else {
       next();
     }
@@ -76,7 +76,7 @@ export function encodeScopePackage(req: $RequestExtend, res: $ResponseExtend, ne
 
 export function expectJson(req: $RequestExtend, res: $ResponseExtend, next: $NextFunctionVer) {
   if (!isObject(req.body)) {
-    return next( ErrorCode.get400('can\'t parse incoming json') );
+    return next( ErrorCode.getBadRequest('can\'t parse incoming json') );
   }
   next();
 }
@@ -87,27 +87,15 @@ export function anti_loop(config: Config) {
       let arr = req.headers.via.split(',');
 
       for (let i=0; i<arr.length; i++) {
-        let m = arr[i].match(/\s*(\S+)\s+(\S+)/);
+        const m = arr[i].match(/\s*(\S+)\s+(\S+)/);
         if (m && m[2] === config.server_id) {
-          return next( ErrorCode.getCode(508, 'loop detected') );
+          return next( ErrorCode.getCode(HTTP_STATUS.LOOP_DETECTED, 'loop detected') );
         }
       }
     }
     next();
   };
 }
-
-/**
- * Express doesn't do etags with requests <= 1024b
- * we use md5 here, it works well on 1k+ bytes, but sucks with fewer data
- * could improve performance using crc32 after benchmarks.
- * @param {Object} data
- * @return {String}
- */
-function md5sum(data) {
-  return crypto.createHash('md5').update(data).digest('hex');
-}
-
 
 export function allow(auth: IAuth) {
   return function(action: string) {
@@ -127,7 +115,7 @@ export function allow(auth: IAuth) {
         } else {
           // last plugin (that's our built-in one) returns either
           // cb(err) or cb(null, true), so this should never happen
-          throw ErrorCode.get500('bug in the auth plugin system');
+          throw ErrorCode.getInternalError('bug in the auth plugin system');
         }
       });
     };
@@ -135,9 +123,9 @@ export function allow(auth: IAuth) {
 }
 
  export function final(body: any, req: $RequestExtend, res: $ResponseExtend, next: $NextFunctionVer) {
-  if (res.statusCode === 401 && !res.getHeader('WWW-Authenticate')) {
+  if (res.statusCode === HTTP_STATUS.UNAUTHORIZED && !res.getHeader('WWW-Authenticate')) {
     // they say it's required for 401, so...
-    res.header('WWW-Authenticate', 'Basic, Bearer');
+    res.header('WWW-Authenticate', `${TOKEN_BASIC}, ${TOKEN_BEARER}`);
   }
 
   try {
@@ -155,7 +143,7 @@ export function allow(auth: IAuth) {
 
       // don't send etags with errors
       if (!res.statusCode || (res.statusCode >= 200 && res.statusCode < 300)) {
-        res.header('ETag', '"' + md5sum(body) + '"');
+        res.header('ETag', '"' + stringToMD5(body) + '"');
       }
     } else {
       // send(null), send(204), etc.
@@ -266,24 +254,24 @@ export function log(req: $RequestExtend, res: $ResponseExtend, next: $NextFuncti
 // Middleware
 export function errorReportingMiddleware(req: $RequestExtend, res: $ResponseExtend, next: $NextFunctionVer) {
   res.report_error = res.report_error || function(err) {
-    if (err.status && err.status >= 400 && err.status < 600) {
+    if (err.status && err.status >= HTTP_STATUS.BAD_REQUEST && err.status < 600) {
       if (_.isNil(res.headersSent) === false) {
         res.status(err.status);
-        next({error: err.message || 'unknown error'});
+        next({error: err.message || API_ERROR.UNKNOWN_ERROR});
       }
     } else {
-      Logger.logger.error( {err: err}
-        , 'unexpected error: @{!err.message}\n@{err.stack}');
+      Logger.logger.error( {err: err}, 'unexpected error: @{!err.message}\n@{err.stack}');
       if (!res.status || !res.send) {
         Logger.logger.error('this is an error in express.js, please report this');
         res.destroy();
       } else if (!res.headersSent) {
-        res.status(500);
-        next({error: 'internal server error'});
+        res.status(HTTP_STATUS.INTERNAL_ERROR);
+        next({error: API_ERROR.INTERNAL_SERVER_ERROR});
       } else {
         // socket should be already closed
       }
     }
   };
+
   next();
 }
