@@ -2,40 +2,39 @@
 
 /* eslint prefer-rest-params: 0 */
 
-import Crypto from 'crypto';
 import assert from 'assert';
-import fs from 'fs';
-import Path from 'path';
 import UrlNode from 'url';
 import _ from 'lodash';
 // $FlowFixMe
-import async from 'async';
-import {ErrorCode, isObject, getLatestVersion, tagVersion, validate_name, semverSort, DIST_TAGS} from './utils';
+import {ErrorCode, isObject, getLatestVersion, tagVersion, validateName, DIST_TAGS} from './utils';
 import {
   generatePackageTemplate, normalizePackage, generateRevision, getLatestReadme, cleanUpReadme,
-  fileExist, noSuchFile, DEFAULT_REVISION, pkgFileName,
+fileExist, noSuchFile, DEFAULT_REVISION, pkgFileName,
 } from './storage-utils';
-import {loadPlugin} from '../lib/plugin-loader';
+import {createTarballHash} from './crypto-utils';
+import {prepareSearchPackage} from './storage-utils';
+import loadPlugin from '../lib/plugin-loader';
 import LocalDatabase from '@verdaccio/local-storage';
 import {UploadTarball, ReadTarball} from '@verdaccio/streams';
 import type {
-  Package,
-  Config,
-  MergeTags,
-  Version,
-  DistFile,
-  Callback,
-  Logger,
+Package,
+Config,
+MergeTags,
+Version,
+DistFile,
+Callback,
+Logger,
 } from '@verdaccio/types';
 import type {
-  ILocalData,
-  IPackageStorage,
+ILocalData,
+IPackageStorage,
 } from '@verdaccio/local-storage';
 import type {
-  IUploadTarball,
-  IReadTarball,
+IUploadTarball,
+IReadTarball,
 } from '@verdaccio/streams';
 import type {IStorage, StringValue} from '../../types';
+import {API_ERROR} from './constants';
 
 /**
  * Implements Storage interface (same for storage.js, local-storage.js, up-storage.js).
@@ -55,12 +54,12 @@ class LocalStorage implements IStorage {
     const storage: any = this._getLocalStorage(name);
 
     if (_.isNil(storage)) {
-      return callback( ErrorCode.get404('this package cannot be added'));
+      return callback( ErrorCode.getNotFound('this package cannot be added'));
     }
 
     storage.createPackage(name, generatePackageTemplate(name), (err) => {
       if (_.isNull(err) === false && err.code === fileExist) {
-        return callback( ErrorCode.get409());
+        return callback( ErrorCode.getConflict());
       }
 
       const latest = getLatestVersion(pkg);
@@ -82,13 +81,13 @@ class LocalStorage implements IStorage {
     let storage: any = this._getLocalStorage(name);
 
     if (_.isNil(storage)) {
-      return callback( ErrorCode.get404());
+      return callback( ErrorCode.getNotFound());
     }
 
     storage.readPackage(name, (err, data) => {
       if (_.isNil(err) === false) {
         if (err.code === noSuchFile) {
-          return callback( ErrorCode.get404());
+          return callback( ErrorCode.getNotFound());
         } else {
           return callback(err);
         }
@@ -99,7 +98,7 @@ class LocalStorage implements IStorage {
       this.localData.remove(name, (removeFailed) => {
         if (removeFailed) {
           // This will happen when database is locked
-          return callback(ErrorCode.get422(removeFailed.message));
+          return callback(ErrorCode.getBadData(removeFailed.message));
         }
 
         storage.deletePackage(pkgFileName, (err) => {
@@ -191,7 +190,7 @@ class LocalStorage implements IStorage {
       }
 
       if (change) {
-        this.logger.debug('updating package info');
+        this.logger.debug({name}, 'updating package @{name} info');
         this._writePackage(name, packageLocalJson, function(err) {
           callback(err, packageLocalJson);
         });
@@ -222,7 +221,7 @@ class LocalStorage implements IStorage {
       metadata = cleanUpReadme(metadata);
 
       if (data.versions[version] != null) {
-        return cb( ErrorCode.get409() );
+        return cb( ErrorCode.getConflict() );
       }
 
       // if uploaded tarball has a different shasum, it's very likely that we have some kind of error
@@ -233,11 +232,17 @@ class LocalStorage implements IStorage {
           if (_.isNil(data._attachments[tarball].shasum) === false && _.isNil(metadata.dist.shasum) === false) {
             if (data._attachments[tarball].shasum != metadata.dist.shasum) {
               const errorMessage = `shasum error, ${data._attachments[tarball].shasum} != ${metadata.dist.shasum}`;
-              return cb( ErrorCode.get400(errorMessage) );
+              return cb( ErrorCode.getBadRequest(errorMessage) );
             }
           }
 
           let currentDate = new Date().toISOString();
+
+          // some old storage do not have this field #740
+          if (_.isNil(data.time)) {
+            data.time = {};
+          }
+
           data.time['modified'] = currentDate;
 
           if (('created' in data.time) === false) {
@@ -254,7 +259,7 @@ class LocalStorage implements IStorage {
 
       this.localData.add(name, (addFailed) => {
         if (addFailed) {
-          return cb(ErrorCode.get422(addFailed.message));
+          return cb(ErrorCode.getBadData(addFailed.message));
         }
 
         cb();
@@ -264,24 +269,25 @@ class LocalStorage implements IStorage {
 
   /**
    * Merge a new list of tags for a local packages with the existing one.
-   * @param {*} name
+   * @param {*} pkgName
    * @param {*} tags
    * @param {*} callback
    */
-  mergeTags(name: string, tags: MergeTags, callback: Callback) {
-    this._updatePackage(name, (data, cb) => {
+  mergeTags(pkgName: string, tags: MergeTags, callback: Callback) {
+    this._updatePackage(pkgName, (data, cb) => {
       /* eslint guard-for-in: 0 */
-      for (let t: string in tags) {
-        if (_.isNull(tags[t])) {
-          delete data[DIST_TAGS][t];
+      for (let tag: string in tags) {
+        // this handle dist-tag rm command
+        if (_.isNull(tags[tag])) {
+          delete data[DIST_TAGS][tag];
           continue;
         }
 
-        if (_.isNil(data.versions[tags[t]])) {
+        if (_.isNil(data.versions[tags[tag]])) {
           return cb( this._getVersionNotFound() );
         }
-        const key: string = tags[t];
-        tagVersion(data, key, t);
+        const version: string = tags[tag];
+        tagVersion(data, version, tag);
       }
       cb();
     }, callback);
@@ -293,7 +299,7 @@ class LocalStorage implements IStorage {
    * @private
    */
   _getVersionNotFound() {
-    return ErrorCode.get404('this version doesn\'t exist');
+    return ErrorCode.getNotFound(API_ERROR.VERSION_NOT_EXIST);
   }
 
   /**
@@ -302,41 +308,42 @@ class LocalStorage implements IStorage {
    * @private
    */
   _getFileNotAvailable() {
-    return ErrorCode.get404('no such file available');
+    return ErrorCode.getNotFound('no such file available');
   }
 
   /**
    * Update the package metadata, tags and attachments (tarballs).
    * Note: Currently supports unpublishing only.
    * @param {*} name
-   * @param {*} pkg
+   * @param {*} incomingPkg
    * @param {*} revision
    * @param {*} callback
    * @return {Function}
    */
   changePackage(name: string,
-                pkg: Package,
+                incomingPkg: Package,
                 revision?: string, callback: Callback) {
-    if (!isObject(pkg.versions) || !isObject(pkg[DIST_TAGS])) {
-      return callback( ErrorCode.get422());
+    if (!isObject(incomingPkg.versions) || !isObject(incomingPkg[DIST_TAGS])) {
+      return callback( ErrorCode.getBadData());
     }
 
-    this._updatePackage(name, (jsonData, cb) => {
-      for (let ver in jsonData.versions) {
-        if (_.isNil(pkg.versions[ver])) {
-          this.logger.info( {name: name, version: ver}, 'unpublishing @{name}@@{version}');
+    this._updatePackage(name, (localData, cb) => {
+      for (let version in localData.versions) {
+        if (_.isNil(incomingPkg.versions[version])) {
+          this.logger.info( {name: name, version: version}, 'unpublishing @{name}@@{version}');
 
-          delete jsonData.versions[ver];
+          delete localData.versions[version];
+          delete localData.time[version];
 
-          for (let file in jsonData._attachments) {
-            if (jsonData._attachments[file].version === ver) {
-              delete jsonData._attachments[file].version;
+          for (let file in localData._attachments) {
+            if (localData._attachments[file].version === version) {
+              delete localData._attachments[file].version;
             }
           }
         }
       }
 
-      jsonData[DIST_TAGS] = pkg[DIST_TAGS];
+      localData[DIST_TAGS] = incomingPkg[DIST_TAGS];
       cb();
     }, function(err) {
       if (err) {
@@ -354,7 +361,7 @@ class LocalStorage implements IStorage {
    */
   removeTarball(name: string, filename: string,
                 revision: string, callback: Callback) {
-    assert(validate_name(filename));
+    assert(validateName(filename));
 
     this._updatePackage(name, (data, cb) => {
       if (data._attachments[filename]) {
@@ -382,16 +389,16 @@ class LocalStorage implements IStorage {
    * @return {Stream}
    */
   addTarball(name: string, filename: string) {
-    assert(validate_name(filename));
+    assert(validateName(filename));
 
     let length = 0;
-    const shaOneHash = Crypto.createHash('sha1');
+    const shaOneHash = createTarballHash();
     const uploadStream: IUploadTarball = new UploadTarball();
     const _transform = uploadStream._transform;
     const storage = this._getLocalStorage(name);
 
-    uploadStream.abort = function() {};
-    uploadStream.done = function() {};
+    (uploadStream: any).abort = function() {};
+    (uploadStream: any).done = function() {};
 
     uploadStream._transform = function(data) {
       shaOneHash.update(data);
@@ -402,7 +409,7 @@ class LocalStorage implements IStorage {
 
     if (name === pkgFileName || name === '__proto__') {
       process.nextTick(() => {
-        uploadStream.emit('error', ErrorCode.get403());
+        uploadStream.emit('error', ErrorCode.getForbidden());
       });
       return uploadStream;
     }
@@ -418,7 +425,7 @@ class LocalStorage implements IStorage {
 
     writeStream.on('error', (err) => {
       if (err.code === fileExist) {
-        uploadStream.emit('error', ErrorCode.get409());
+        uploadStream.emit('error', ErrorCode.getConflict());
       } else if (err.code === noSuchFile) {
         // check if package exists to throw an appropriate message
         this.getPackageMetadata(name, function(_err, res) {
@@ -453,13 +460,13 @@ class LocalStorage implements IStorage {
       });
     });
 
-    uploadStream.abort = function() {
+    (uploadStream: any).abort = function() {
       writeStream.abort();
     };
 
-    uploadStream.done = function() {
+    (uploadStream: any).done = function() {
       if (!length) {
-        uploadStream.emit('error', ErrorCode.get422('refusing to accept zero-length file'));
+        uploadStream.emit('error', ErrorCode.getBadData('refusing to accept zero-length file'));
         writeStream.abort();
       } else {
         writeStream.done();
@@ -478,7 +485,7 @@ class LocalStorage implements IStorage {
    * @return {ReadTarball}
    */
   getTarball(name: string, filename: string): IReadTarball {
-    assert(validate_name(filename));
+    assert(validateName(filename));
 
     const storage: IPackageStorage = this._getLocalStorage(name);
 
@@ -513,9 +520,9 @@ class LocalStorage implements IStorage {
   _streamSuccessReadTarBall(storage: any, filename: string): IReadTarball {
     const stream: IReadTarball = new ReadTarball();
     const readTarballStream = storage.readTarball(filename);
-    const e404 = ErrorCode.get404;
+    const e404 = ErrorCode.getNotFound;
 
-    stream.abort = function() {
+    (stream: any).abort = function() {
       if (_.isNil(readTarballStream) === false) {
         readTarballStream.abort();
       }
@@ -551,7 +558,7 @@ class LocalStorage implements IStorage {
   getPackageMetadata(name: string, callback?: Callback = () => {}): void {
     const storage: IPackageStorage = this._getLocalStorage(name);
     if (_.isNil(storage)) {
-      return callback( ErrorCode.get404() );
+      return callback( ErrorCode.getNotFound() );
     }
 
     this._readPackage(name, storage, callback);
@@ -564,53 +571,24 @@ class LocalStorage implements IStorage {
    * @return {Function}
    */
   search(startKey: string, options: any) {
-    const stream = new UploadTarball({objectMode: true});
+    const stream = new ReadTarball({objectMode: true});
 
-    this._eachPackage((item, cb) => {
-      fs.stat(item.path, (err, stats) => {
-        if (_.isNil(err) === false) {
-          return cb(err);
-        }
-
-        if (stats.mtime.getTime() > parseInt(startKey, 10)) {
-          this.getPackageMetadata(item.name, (err: Error, data: Package) => {
-            if (err) {
-              return cb(err);
-            }
-
-            const listVersions: Array<string> = Object.keys(data.versions);
-            const versions: Array<string> = semverSort(listVersions);
-            const latest: string = data[DIST_TAGS] && data[DIST_TAGS].latest ? data[DIST_TAGS].latest : versions.pop();
-
-            if (data.versions[latest]) {
-              const version: Version = data.versions[latest];
-              const pkg: any = {
-                name: version.name,
-                description: version.description,
-                'dist-tags': {latest},
-                maintainers: version.maintainers || [version.author].filter(Boolean),
-                author: version.author,
-                repository: version.repository,
-                readmeFilename: version.readmeFilename || '',
-                homepage: version.homepage,
-                keywords: version.keywords,
-                bugs: version.bugs,
-                license: version.license,
-                time: {
-                  modified: item.time ? new Date(item.time).toISOString() : stats.mtime,
-                },
-                versions: {[latest]: 'latest'},
-              };
-
-              stream.push(pkg);
-            }
-
-            cb();
-          });
-        } else {
+    this._searchEachPackage((item, cb) => {
+      if (item.time > parseInt(startKey, 10)) {
+        this.getPackageMetadata(item.name, (err: Error, data: Package) => {
+          if (err) {
+            return cb(err);
+          }
+          const time = new Date(item.time).toISOString();
+          const result = prepareSearchPackage(data, time);
+          if (_.isNil(result) === false) {
+            stream.push(result);
+          }
           cb();
-        }
-      });
+        });
+      } else {
+        cb();
+      }
     }, function onEnd(err) {
       if (err) {
         return stream.emit('error', err);
@@ -623,11 +601,11 @@ class LocalStorage implements IStorage {
 
   /**
    * Retrieve a wrapper that provide access to the package location.
-   * @param {Object} packageInfo package name.
+   * @param {Object} pkgName package name.
    * @return {Object}
    */
-  _getLocalStorage(packageInfo: string): IPackageStorage {
-    return this.localData.getPackageStorage(packageInfo);
+  _getLocalStorage(pkgName: string): IPackageStorage {
+    return this.localData.getPackageStorage(pkgName);
   }
 
   /**
@@ -639,7 +617,7 @@ class LocalStorage implements IStorage {
     storage.readPackage(name, (err, result) => {
       if (err) {
         if (err.code === noSuchFile) {
-          return callback( ErrorCode.get404() );
+          return callback( ErrorCode.getNotFound() );
         } else {
           return callback(this._internalError(err, pkgFileName, 'error reading'));
         }
@@ -649,79 +627,61 @@ class LocalStorage implements IStorage {
     });
   }
 
+  _getCustomPackageLocalStorages() {
+    const storages = {};
+
+    // add custom storage if exist
+    if (this.config.storage) {
+      storages[this.config.storage] = true;
+    }
+
+    const {packages} = this.config;
+
+    if (packages) {
+      const listPackagesConf = Object.keys(packages || {});
+
+      listPackagesConf.map( (pkg) => {
+        if (packages[pkg].storage) {
+          storages[packages[pkg].storage] = true;
+        }
+      });
+    }
+
+    return storages;
+  }
+
   /**
    * Walks through each package and calls `on_package` on them.
    * @param {*} onPackage
    * @param {*} onEnd
    */
-  _eachPackage(onPackage: Callback, onEnd: Callback) {
-    const storages = {};
-
-    storages[this.config.storage] = true;
-    if (this.config.packages) {
-      Object.keys(this.config.packages || {}).map( (pkg) => {
-        if (this.config.packages[pkg].storage) {
-          storages[this.config.packages[pkg].storage] = true;
-        }
-      });
+  _searchEachPackage(onPackage: Callback, onEnd: Callback) {
+    // save wait whether plugin still do not support search functionality
+    if (_.isNil(this.localData.search)) {
+      this.logger.warn('plugin search not implemented yet');
+      onEnd();
+    } else {
+      this.localData.search(onPackage, onEnd, validateName);
     }
-    const base = Path.dirname(this.config.self_path);
-
-    async.eachSeries(Object.keys(storages), function(storage, cb) {
-      fs.readdir(Path.resolve(base, storage), function(err, files) {
-        if (err) {
-          return cb(err);
-        }
-
-        async.eachSeries(files, function(file, cb) {
-          if (file.match(/^@/)) {
-            // scoped
-            fs.readdir(Path.resolve(base, storage, file), function(err, files) {
-              if (err) {
-                return cb(err);
-              }
-
-              async.eachSeries(files, (file2, cb) => {
-                if (validate_name(file2)) {
-                  onPackage({
-                    name: `${file}/${file2}`,
-                    path: Path.resolve(base, storage, file, file2),
-                  }, cb);
-                } else {
-                  cb();
-                }
-              }, cb);
-            });
-          } else if (validate_name(file)) {
-            onPackage({
-              name: file,
-              path: Path.resolve(base, storage, file),
-            }, cb);
-          } else {
-            cb();
-          }
-        }, cb);
-      });
-    }, onEnd);
   }
 
   /**
    * Retrieve either a previous created local package or a boilerplate.
-   * @param {*} name
+   * @param {*} pkgName
    * @param {*} callback
    * @return {Function}
    */
-  _readCreatePackage(name: string, callback: Callback) {
-    const storage: any = this._getLocalStorage(name);
+  _readCreatePackage(pkgName: string, callback: Callback) {
+    const storage: any = this._getLocalStorage(pkgName);
     if (_.isNil(storage)) {
-      return this._createNewPackage(name, callback);
+      return this._createNewPackage(pkgName, callback);
     }
 
-    storage.readPackage(name, (err, data) => {
+    storage.readPackage(pkgName, (err, data) => {
       // TODO: race condition
       if (_.isNil(err) === false) {
         if (err.code === noSuchFile) {
-          data = generatePackageTemplate(name);
+          data = generatePackageTemplate(pkgName);
         } else {
           return callback(this._internalError(err, pkgFileName, 'error reading'));
         }
@@ -745,7 +705,7 @@ class LocalStorage implements IStorage {
   _internalError(err: string, file: string, message: string) {
     this.logger.error( {err: err, file: file}, `${message}  @{file}: @{!err.message}` );
 
-    return ErrorCode.get500();
+    return ErrorCode.getInternalError();
   }
 
   /**
@@ -758,7 +718,7 @@ class LocalStorage implements IStorage {
     const storage: IPackageStorage = this._getLocalStorage(name);
 
     if (!storage) {
-      return callback( ErrorCode.get404() );
+      return callback( ErrorCode.getNotFound() );
     }
 
     storage.updatePackage(name, updateHandler, this._writePackage.bind(this), normalizePackage,
@@ -786,7 +746,10 @@ class LocalStorage implements IStorage {
       json._rev = DEFAULT_REVISION;
     }
 
-    json._rev = generateRevision(json._rev);
+    // this is intended in debug mode we do not want modify the store revision
+    if (_.isNil(this.config._debug)) {
+      json._rev = generateRevision(json._rev);
+    }
 
     return json;
   }
@@ -838,7 +801,7 @@ class LocalStorage implements IStorage {
     return this.localData.setSecret(config.checkSecretKey(secretKey));
   }
 
-  _loadStorage(config: Config, logger: Logger) {
+  _loadStorage(config: Config, logger: Logger): ILocalData {
     const Storage = this._loadStorePlugin();
 
     if (_.isNil(Storage)) {
@@ -849,13 +812,13 @@ class LocalStorage implements IStorage {
     }
   }
 
-  _loadStorePlugin() {
+  _loadStorePlugin(): ILocalData {
     const plugin_params = {
       config: this.config,
       logger: this.logger,
     };
 
-    return _.head(loadPlugin(this.config, this.config.store, plugin_params, function(plugin) {
+    return _.head(loadPlugin(this.config, this.config.store, plugin_params, (plugin) => {
       return plugin.getPackageStorage;
     }));
   }
