@@ -1,29 +1,29 @@
 // @flow
 
 import _ from 'lodash';
-import crypto from 'crypto';
-import {ErrorCode, isObject, normalizeDistTags, DIST_TAGS} from './utils';
+import {ErrorCode, isObject, normalizeDistTags, DIST_TAGS, semverSort} from './utils';
 import Search from './search';
+import {generateRandomHexString} from '../lib/crypto-utils';
 
 import type {Package, Version} from '@verdaccio/types';
 import type {IStorage} from '../../types';
+import {API_ERROR, HTTP_STATUS} from './constants';
 
 const pkgFileName = 'package.json';
 const fileExist: string = 'EEXISTS';
 const noSuchFile: string = 'ENOENT';
-const resourceNotAvailable: string = 'EAGAIN';
-const DEFAULT_REVISION: string = `0-0000000000000000`;
+export const DEFAULT_REVISION: string = `0-0000000000000000`;
 
 const generatePackageTemplate = function(name: string): Package {
   return {
     // standard things
     name,
     versions: {},
-    'dist-tags': {},
     time: {},
+    [DIST_TAGS]: {},
+    _uplinks: {},
     _distfiles: {},
     _attachments: {},
-    _uplinks: {},
     _rev: '',
   };
 };
@@ -42,7 +42,9 @@ function normalizePackage(pkg: Package) {
     'time'];
 
   pkgProperties.forEach((key) => {
-    if (_.isNil(isObject(pkg[key]))) {
+    const pkgProp = pkg[key];
+
+    if (_.isNil(pkgProp) || isObject(pkgProp) === false) {
       pkg[key] = {};
     }
   });
@@ -60,7 +62,7 @@ function normalizePackage(pkg: Package) {
 function generateRevision(rev: string): string {
   const _rev = rev.split('-');
 
-  return ((+_rev[0] || 0) + 1) + '-' + crypto.pseudoRandomBytes(8).toString('hex');
+  return ((+_rev[0] || 0) + 1) + '-' + generateRandomHexString();
 }
 
 function getLatestReadme(pkg: Package): string {
@@ -122,11 +124,11 @@ export function cleanUpLinksRef(keepUpLinkData: boolean, result: Package): Packa
 export function checkPackageLocal(name: string, localStorage: IStorage): Promise<any> {
   return new Promise((resolve, reject) => {
     localStorage.getPackageMetadata(name, (err, results) => {
-      if (!_.isNil(err) && err.status !== 404) {
+      if (!_.isNil(err) && err.status !== HTTP_STATUS.NOT_FOUND) {
         return reject(err);
       }
       if (results) {
-        return reject(ErrorCode.get409('this package is already present'));
+        return reject(ErrorCode.getConflict(API_ERROR.PACKAGE_EXIST));
       }
       return resolve();
     });
@@ -151,25 +153,25 @@ export function checkPackageRemote(name: string, isAllowPublishOffline: boolean,
     // $FlowFixMe
     syncMetadata(name, null, {}, (err, packageJsonLocal, upLinksErrors) => {
       // something weird
-      if (err && err.status !== 404) {
+      if (err && err.status !== HTTP_STATUS.NOT_FOUND) {
         return reject(err);
       }
 
       // checking package exist already
       if (_.isNil(packageJsonLocal) === false) {
-        return reject(ErrorCode.get409('this package is already present'));
+        return reject(ErrorCode.getConflict(API_ERROR.PACKAGE_EXIST));
       }
 
       for (let errorItem = 0; errorItem < upLinksErrors.length; errorItem++) {
         // checking error
         // if uplink fails with a status other than 404, we report failure
         if (_.isNil(upLinksErrors[errorItem][0]) === false) {
-          if (upLinksErrors[errorItem][0].status !== 404) {
+          if (upLinksErrors[errorItem][0].status !== HTTP_STATUS.NOT_FOUND) {
             if (isAllowPublishOffline) {
               return resolve();
             }
 
-            return reject(ErrorCode.get503('one of the uplinks is down, refuse to publish'));
+            return reject(ErrorCode.getServiceUnavailable(API_ERROR.UPLINK_OFFLINE_PUBLISH));
           }
         }
       }
@@ -179,16 +181,51 @@ export function checkPackageRemote(name: string, isAllowPublishOffline: boolean,
   });
 }
 
+export function mergeUplinkTimeIntoLocal(localMetadata: Package, remoteMetadata: Package): any {
+  if ('time' in remoteMetadata) {
+    return Object.assign({}, localMetadata.time, remoteMetadata.time);
+  }
+
+  return localMetadata.time;
+}
+
+export function prepareSearchPackage(data: Package, time: mixed) {
+  const listVersions: Array<string> = Object.keys(data.versions);
+  const versions: Array<string> = semverSort(listVersions);
+  const latest: string = data[DIST_TAGS] && data[DIST_TAGS].latest ? data[DIST_TAGS].latest : versions.pop();
+
+  if (data.versions[latest]) {
+    const version: Version = data.versions[latest];
+    const pkg: any = {
+      name: version.name,
+      description: version.description,
+      [DIST_TAGS]: {latest},
+      maintainers: version.maintainers || [version.author].filter(Boolean),
+      author: version.author,
+      repository: version.repository,
+      readmeFilename: version.readmeFilename || '',
+      homepage: version.homepage,
+      keywords: version.keywords,
+      bugs: version.bugs,
+      license: version.license,
+      time: {
+        modified: time,
+      },
+      versions: {[latest]: 'latest'},
+    };
+
+    return pkg;
+  }
+}
+
 export {
   generatePackageTemplate,
   normalizePackage,
   generateRevision,
   getLatestReadme,
   cleanUpReadme,
-  DEFAULT_REVISION,
   fileExist,
   noSuchFile,
   pkgFileName,
-  resourceNotAvailable,
 };
 
