@@ -17,17 +17,42 @@ class API {
         url = window.VERDACCIO_API_URL + url;
       }
 
-      function handleErrors(response) {
-        if (!response.ok) {
-            throw Error(response.statusText);
+      /**
+       * Handles response according to content type
+       * @param {object} response
+       * @returns {promise}
+       */
+      function handleResponseType(response) {
+        if (response.headers) {
+          const contentType = response.headers.get('Content-Type');
+          if (contentType.includes('application/pdf')) {
+            return Promise.all([response.ok, response.blob()]);
+          }
+          if (contentType.includes('application/json')) {
+            return Promise.all([response.ok, response.json()]);
+          }
+          // it includes all text types
+          if (contentType.includes('text/')) {
+            return Promise.all([response.ok, response.text()]);
+          }
         }
-        return response;
       }
 
-      return fetch(url, {
-        method,
-        ...options
-      }).then(handleErrors);
+      return new Promise((resolve, reject) => {
+        fetch(url, {
+          method,
+          credentials: 'same-origin',
+          ...options
+        })
+        .then(handleResponseType)
+        .then(([responseOk, body]) => {
+          if (responseOk) {
+            resolve(body);
+          } else {
+            reject(body);
+          }
+        });
+      });
     }
 }
 
