@@ -1,3 +1,3 @@
-require("babel-polyfill");
+require('babel-polyfill');
 require('babel-register');
 module.exports = require('./setup');
