@@ -1,3 +1,5 @@
+import {API_ERROR, HTTP_STATUS} from "../../../src/lib/constants";
+
 export default function(server) {
   describe('npm adduser', () => {
     const user = String(Math.random());
@@ -5,7 +7,7 @@ export default function(server) {
 
     beforeAll(function() {
       return server.auth(user, pass)
-               .status(201)
+               .status(HTTP_STATUS.CREATED)
                .body_ok(/user .* created/);
     });
 
@@ -13,14 +15,14 @@ export default function(server) {
 
     test('should log in', () => {
       return server.auth(user, pass)
-               .status(201)
+               .status(HTTP_STATUS.CREATED)
                .body_ok(/you are authenticated as/);
     });
 
     test('should not register more users', () => {
       return server.auth(String(Math.random()), String(Math.random()))
-               .status(409)
-               .body_error(/maximum amount of users reached/);
+               .status(HTTP_STATUS.CONFLICT)
+               .body_error(API_ERROR.MAX_USERS_REACHED);
     });
   });
 }
