@@ -1,8 +1,10 @@
+import {CREDENTIALS} from "../config.functional";
+
 module.exports = function(server) {
 
     test('who am I?', () => {
       return server.whoami().then(function (username) {
-        expect(username).toMatch('test');
+        expect(username).toMatch(CREDENTIALS.user);
       });
     });
 
