@@ -5,48 +5,54 @@ const config = {
   storage: path.join(__dirname, '../store/test-storage'),
   uplinks: {
     'npmjs': {
-      'url': 'https://registry.npmjs.org/'
+      'url': 'http://localhost:4873/'
     }
   },
   packages: {
     '@*/*': {
-      allow_access: '$all',
-      allow_publish: '$all',
+      access: '$all',
+      publish: '$all',
       proxy: 'npmjs'
     },
 
     'forbidden-place': {
-      allow_access: 'nobody',
-      allow_publish: '$all'
+      access: 'nobody',
+      publish: '$all'
     },
 
     'react': {
-      allow_access: '$all',
-      allow_publish: '$all',
+      access: '$all',
+      publish: '$all',
+      proxy: 'npmjs'
+    },
+
+    'corrupted-package': {
+      access: '$all',
+      publish: '$all',
       proxy: 'npmjs'
     },
 
     'jquery': {
-      allow_access: '$all',
-      allow_publish: '$all',
+      access: '$all',
+      publish: '$all',
       proxy: 'npmjs'
     },
     'auth-package': {
-      allow_access: '$authenticated',
-      allow_publish: '$authenticated'
+      access: '$authenticated',
+      publish: '$authenticated'
     },
     'vue': {
-      allow_access: '$authenticated',
-      allow_publish: '$authenticated',
+      access: '$authenticated',
+      publish: '$authenticated',
       proxy: 'npmjs'
     },
     '*': {
-      allow_access: '$all',
-      allow_publish: '$all'
+      access: '$all',
+      publish: '$all'
     },
   },
   logs: [
-    {type: 'stdout', format: 'pretty', level: 'fatal'},
+    {type: 'stdout', format: 'pretty', level: 'warn'},
   ],
 };
 
