@@ -1,20 +1,77 @@
-/**
- * Copyright (c) 2017-present, Facebook, Inc.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
 const React = require('react');
 
 const CompLibrary = require('../../core/CompLibrary.js');
 const translate = require("../../server/translate.js").translate;
 const translation = require('../../server/translation.js');
-const MarkdownBlock = CompLibrary.MarkdownBlock; /* Used to read markdown */
 const Container = CompLibrary.Container;
 const GridBlock = CompLibrary.GridBlock;
 
 const siteConfig = require(process.cwd() + '/siteConfig.js');
+
+class TweetsSection extends React.Component {
+  render() {
+    return(
+      <div className="tweets-section">
+        <div>
+          {this.renderTweetUpRight()}
+        </div>
+        <div>
+          {this.renderTweetUpLeft()}
+        </div>
+        <div>
+          {this.renderTweetDownRight()}
+        </div>
+        <div>
+          {this.renderTweetDownLeft()}
+        </div>
+      </div>);
+  }
+
+  renderTweetUpRight() {
+    return (<blockquote className="twitter-tweet" data-cards="hidden" data-lang="es"><p lang="en" dir="ltr">In a world
+        where <a href="https://twitter.com/npmjs?ref_src=twsrc%5Etfw">@npmjs</a> and other core infrastructure fully
+        embraced the &quot;move fast and break things&quot; mindset, it&#39;s great to see other people care deeply about
+        stability. <a href="https://t.co/uVk7xFeiwU">https://t.co/uVk7xFeiwU</a> &quot;just works&quot; and has saved us
+        countless headaches. Thanks <a href="https://twitter.com/verdaccio_npm?ref_src=twsrc%5Etfw">@verdaccio_npm</a> !
+      </p>&mdash; Sheet JS (@SheetJS) <a
+        href="https://twitter.com/SheetJS/status/1002609907370250241?ref_src=twsrc%5Etfw">1 de junio de 2018</a>
+      </blockquote>);
+  }
+
+  renderTweetUpLeft() {
+    return (
+      <blockquote className="twitter-tweet" data-lang="es"><p lang="en" dir="ltr">Woke up to <a
+        href="https://twitter.com/timer150?ref_src=twsrc%5Etfw">@timer150</a> fixing end-to-end test flakiness in Create
+        React App 👏 Background: we have a Lerna monorepo and used very complex hacks for integration testing of
+        generated projects. Solution: run a local npm registry to simulate a publish 😁 <a
+          href="https://t.co/ggNfS4PpFv">https://t.co/ggNfS4PpFv</a></p>&mdash; Dan Abramov (@dan_abramov) <a
+        href="https://twitter.com/dan_abramov/status/951427300070916096?ref_src=twsrc%5Etfw">11 de enero de 2018</a>
+      </blockquote>);
+  }
+
+  renderTweetDownRight() {
+    return (
+      <blockquote className="twitter-tweet" data-lang="es"><p lang="en" dir="ltr">Verdaccio is my personal hero. A
+        lightweight npm registry &amp; cache. Just two commands to install + use it. <a
+          href="https://twitter.com/hashtag/incredibile?src=hash&amp;ref_src=twsrc%5Etfw">#incredibile</a> <a
+          href="https://t.co/X0uNgS6UMz">https://t.co/X0uNgS6UMz</a></p>&mdash; Manfred Steyer (@ManfredSteyer) <a
+        href="https://twitter.com/ManfredSteyer/status/1002153128140136448?ref_src=twsrc%5Etfw">31 de mayo de 2018</a>
+      </blockquote>)
+  }
+
+  renderTweetDownLeft() {
+    return (
+      <blockquote className="twitter-tweet" data-conversation="none" data-cards="hidden" data-lang="es">
+        <p lang="en" dir="ltr">
+          <a href="https://twitter.com/verdaccio_npm?ref_src=twsrc%5Etfw">@verdaccio_npm</a> is a pretty good open-source
+        tool and seems to be growing. We&#39;re currently using them at NodeSource for some internal caching and
+        modules, and are happy so far. Easy to setup for both business and use.<br/><br/>🔗
+          <a href="https://t.co/ow5JRgZYrU">https://t.co/ow5JRgZYrU</a>
+        </p>&mdash; Tierney Cyren (@bitandbang) <a
+        href="https://twitter.com/bitandbang/status/1001297542779424768?ref_src=twsrc%5Etfw">29 de mayo de 2018</a>
+      </blockquote>);
+  }
+}
 
 class Button extends React.Component {
   render() {
@@ -32,6 +89,32 @@ Button.defaultProps = {
   target: '_self',
 };
 
+class HeaderButtons extends React.Component {
+  buildUrl(relativePath) {
+    return `${siteConfig.baseUrl}docs/${this.props.language}${relativePath}`;
+  }
+
+  render() {
+    return (
+      <div className="pluginRowBlock">
+        <Button href={this.buildUrl('/what-is-verdaccio.html')}>
+          <translate>
+            Get Started
+          </translate>
+        </Button>
+        <Button href="https://github.com/verdaccio">
+          Github
+        </Button>
+        <Button href={this.buildUrl('/contributing.html')}>
+          <translate>
+            Contribute
+          </translate>
+        </Button>
+      </div>
+    )
+  }
+}
+
 class HomeSplash extends React.Component {
   render() {
     return (
@@ -39,57 +122,27 @@ class HomeSplash extends React.Component {
         <div className="homeSplashFade">
           <div className="wrapper homeWrapper">
             <div className="inner">
-              <h2 className="projectTitle">
-              <img title={siteConfig.title} src={siteConfig.baseUrl + 'img/verdaccio@2x.png'} />
+              <h1 className="projectTitle">
+              <img title={siteConfig.title} alt={siteConfig.title} src={siteConfig.baseUrl + 'img/verdaccio@2x.png'} />
                 <small>
                   {
-                    translation[this.props.language]['localized-strings']
-                      .tagline
+                    translation[this.props.language]['localized-strings'].tagline
                   }
                 </small>
-              </h2>
+              </h1>
               <div className="section promoSection">
                 <div className="promoRow">
-                  <div className="pluginRowBlock">
-                    <Button
-                      href={
-                        siteConfig.baseUrl +
-                        'docs/' +
-                        this.props.language +
-                        '/what-is-verdaccio.html'
-                      }>
-                      <translate>
-                        Get Started
-                      </translate>
-                    </Button>
-                    <Button href="https://github.com/verdaccio">Github</Button>
-                    <Button
-                      href={
-                        siteConfig.baseUrl +
-                        'docs/' +
-                        this.props.language +
-                        '/contributing.html'
-                      }>
-                      <translate>
-                        Contribute
-                      </translate>
-                    </Button>
-                  </div>
+                  <HeaderButtons language={this.props.language}/>
                 </div>
               </div>
-              <div className="githubButton" style={{minHeight: '20px'}}>
-                <a
-                  className="github-button"
-                  href={siteConfig.repoUrl}
-                  data-icon="octicon-star"
-                  data-count-href="/verdaccio/verdaccio/stargazers"
-                  data-show-count={true}
-                  data-count-aria-label="# stargazers on GitHub"
-                  aria-label="Star verdaccio/verdaccio on GitHub"
-                >
-                  Star
-                </a>
-              </div>
+              <iframe
+                src={"https://ghbtns.com/github-btn.html?user=verdaccio&repo=verdaccio&type=star&count=true&size=large"}
+                frameBorder="0"
+                scrolling="0"
+                width="160px"
+                height="30px"
+                style={{marginTop: '8px'}}
+              />
             </div>
           </div>
         </div>
@@ -98,6 +151,32 @@ class HomeSplash extends React.Component {
   }
 }
 
+const FeatureSection = () => (
+  <GridBlock
+    className="feature-section"
+    align="center"
+    contents={[
+      {
+        content: '```bash\n' + 'npm install --global verdaccio'
+      }
+    ]}
+    layout="twoColumn"
+  />
+);
+
+const MiddleTitle = (props) => {
+  return (
+    <div className="productShowcaseSection paddingBottom"
+         style={{textAlign: 'center'}}>
+      <h2>
+        {props.children}
+      </h2>
+    </div>
+  );
+}
+
+
+
 class Index extends React.Component {
   render() {
     let language = this.props.language || 'en';
@@ -105,10 +184,10 @@ class Index extends React.Component {
       .filter(user => {
         return user.pinned;
       })
-      .map(user => {
+      .map((user, i) => {
         return (
-          <a href={user.infoLink}>
-            <img src={user.image} title={user.caption} />
+          <a href={user.infoLink} key={i} target="_blank" rel="noopener">
+            <img src={user.image} alt={user.caption} title={user.caption} />
           </a>
         );
       });
@@ -117,51 +196,22 @@ class Index extends React.Component {
       <div>
         <HomeSplash language={language} />
         <div className="mainContainer">
-          <Container padding={['bottom', 'top']}>
-            <GridBlock
-              align="center"
-              contents={[
-                {
-                  content: '```bash\n' +
-                  '$> npm install --global verdaccio \n' +
-                  '$> yarn global add verdaccio\n' +
-                  '```',
-                  title: <translate>Easy to Install</translate>,
-                },
-                {
-                  content: '```bash\n' +
-                  '$> npm set registry http://localhost:4873 \n' +
-                  '$> npm adduser --registry http://localhost:4873\n' +
-                  '```',
-                  title: <translate>Easy to Set Up</translate>,
-                }
-              ]}
-              layout="fourColumn"
-            />
+          <Container padding={['top']}>
+            <FeatureSection/>
           </Container>
-
-          <div
-            className="productShowcaseSection paddingBottom"
-            style={{textAlign: 'center'}}>
-            <h2>
-              <translate>
-                Easy to Use
-              </translate>
-            </h2>
-            <MarkdownBlock>
-              ```bash&#8232;
-                $> verdaccio &#8232;
-                warn --- config file  - /home/.config/verdaccio/config.yaml &#8232;
-                warn --- http address - http://localhost:4873/ - verdaccio/3.0.0&#8232;
-              ```
-            </MarkdownBlock>
-            <h2>
+          <MiddleTitle>
               <translate>
                 That’s it ! Enjoy your private package manager.
               </translate>
-            </h2>
-          </div>
-
+          </MiddleTitle>
+          <Container>
+            <TweetsSection/>
+          </Container>
+          <MiddleTitle>
+            <translate>
+              Many great developers are already enjoying Verdaccio, join the community!
+            </translate>
+          </MiddleTitle>
           <Container padding={['bottom', 'top']} background="light">
             <GridBlock
               contents={[
@@ -169,6 +219,7 @@ class Index extends React.Component {
                   content: <translate>**npm**, **yarn** and **pnpm** are part of any development workflow
                     we try to catch up with the latest updates.</translate>,
                   image: siteConfig.baseUrl + 'img/package_managers_grey.png',
+                  imageAlt: 'The most popular npm clients are supported',
                   imageAlign: 'right',
                   title: <translate>The most popular npm clients are supported</translate>,
                 },
