@@ -17,8 +17,8 @@ class Users extends React.Component {
   render() {
     const showcase = siteConfig.users.map((user, i) => {
       return (
-        <a href={user.infoLink} key={i}>
-          <img src={user.image} title={user.caption} />
+        <a href={user.infoLink} key={i} target="_blank" rel="noopener">
+          <img src={user.image} alt={user.caption} title={user.caption} />
         </a>
       );
     });
